Rename router → routes
