+++ server/src/routes/admin/thing/thing.controller.ts
