+++ server/src/routes/api/thing/index.ts
